refactor: annex into program
