+++ jac/jaclang/compiler/passes/main/import_pass.py
@@ -15,7 +15,6 @@ import jaclang.compiler.absyntree as ast
 from jaclang.compiler.passes import AstPass
 from jaclang.compiler.passes.main import DefUsePass, SymTabBuildPass
 from jaclang.compiler.passes.main.sym_tab_build_pass import PyInspectSymTabBuildPass
-from jaclang.settings import settings
 from jaclang.utils.log import logging
 
 
@@ -30,7 +29,6 @@ class JacImportPass(AstPass):
     def enter_module(self, node: ast.Module) -> None:
         """Run Importer."""
         self.cur_node = node
-        self.annex_impl(node)
         self.terminate()  # Turns off auto traversal for deliberate traversal
         all_imports = self.get_all_sub_nodes(node, ast.ModulePath)
         for i in all_imports:
@@ -78,53 +76,6 @@ class JacImportPass(AstPass):
             self.prog.last_imported.append(mod)
 
     # TODO: Refactor this to a function for impl and function for test
-    def annex_impl(self, node: ast.Module) -> None:
-        """Annex impl and test modules."""
-        if node.stub_only:
-            return
-        if not node.loc.mod_path:
-            self.log_error("Module has no path")
-        if not node.loc.mod_path.endswith(".jac"):
-            return
-        base_path = node.loc.mod_path[:-4]
-        directory = os.path.dirname(node.loc.mod_path)
-        if not directory:
-            directory = os.getcwd()
-            base_path = os.path.join(directory, base_path)
-        impl_folder = base_path + ".impl"
-        test_folder = base_path + ".test"
-        search_files = [
-            os.path.join(directory, impl_file) for impl_file in os.listdir(directory)
-        ]
-        if os.path.exists(impl_folder):
-            search_files += [
-                os.path.join(impl_folder, impl_file)
-                for impl_file in os.listdir(impl_folder)
-            ]
-        if os.path.exists(test_folder):
-            search_files += [
-                os.path.join(test_folder, test_file)
-                for test_file in os.listdir(test_folder)
-            ]
-        for cur_file in search_files:
-            if node.loc.mod_path.endswith(cur_file):
-                continue
-            if (
-                cur_file.startswith(f"{base_path}.")
-                or impl_folder == os.path.dirname(cur_file)
-            ) and cur_file.endswith(".impl.jac"):
-                mod = self.prog.jac_file_to_pass(file_path=cur_file, schedule=[]).ir_out
-                if mod:
-                    node.add_kids_left(mod.kid, parent_update=True, pos_update=False)
-                    node.impl_mod.append(mod)
-            if (
-                cur_file.startswith(f"{base_path}.")
-                or test_folder == os.path.dirname(cur_file)
-            ) and cur_file.endswith(".test.jac"):
-                mod = self.prog.jac_file_to_pass(file_path=cur_file, schedule=[]).ir_out
-                if mod and not settings.ignore_test_annex:
-                    node.test_mod.append(mod)
-                    node.add_kids_right(mod.kid, parent_update=True, pos_update=False)
 
     def import_jac_mod_from_dir(self, target: str) -> ast.Module | None:
         """Import a module from a directory."""
